+++ ReleaseNotes.md
@@ -18,3 +18,4 @@
 - Add MileEnableImmersiveDarkModeForWindow function.
 - Add MILE_TOKEN_PRIVILEGE_LUID_CLASS enum.
 - Add MileAllocateMemory function.
+- Add MileReallocateMemory function.